requirements.txt
# Core ML and Audio Processing
torch>=2.0.0
torchaudio>=2.0.0
librosa>=0.10.0
numpy>=1.24.0
pandas>=2.0.0
soundfile>=0.12.0
scipy>=1.10.0

# Configuration and Utilities
pyyaml>=6.0
omegaconf>=2.3.0
tqdm>=4.65.0
matplotlib>=3.7.0
plotly>=5.15.0

# Audio Enhancement and Evaluation
pesq>=0.0.4
pystoi>=0.3.3
asteroid>=0.6.0

# Machine Learning
scikit-learn>=1.3.0
transformers>=4.30.0

# Development and Testing
black>=23.0.0
ruff>=0.0.280
pytest>=7.4.0
pre-commit>=3.3.0

# Demo and Visualization
streamlit>=1.25.0
gradio>=3.40.0

# Optional: Advanced Features
# wandb>=0.15.0  # For experiment tracking
# fastapi>=0.100.0  # For API serving
# uvicorn>=0.23.0  # For API serving
